web/backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# WebSocket support
websockets>=12.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CORS and security
python-jose[cryptography]>=3.3.0

# Async support
aiofiles>=23.2.1
httpx>=0.25.0

# Image/video processing
pillow>=10.1.0
numpy>=1.24.0

# For connecting to Atlas
opencv-python-headless>=4.8.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
